moved wordset class to the wordset file
