+++ frontend/src/ts/utils/misc.ts
@@ -1409,23 +1409,6 @@ export function memoizeAsync<T extends (...args: any) => Promise<any>>(
   }) as T;
 }
 
-export class Wordset {
-  public words: string[];
-  public length: number;
-  constructor(words: string[]) {
-    this.words = words;
-    this.length = this.words.length;
-  }
-
-  public randomWord(mode: MonkeyTypes.FunboxWordsFrequency): string {
-    if (mode === "zipf") {
-      return this.words[dreymarIndex(this.words.length)];
-    } else {
-      return randomElementFromArray(this.words);
-    }
-  }
-}
-
 export class Section {
   public title: string;
   public author: string;
